pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-refinery"
version = "0.1.0"
description = "Document Intelligence Refinery — structured extraction and chunking pipeline (docling + OpenRouter vision)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "pydantic>=2.6,<3",
    "pyyaml>=6.0",
    "python-magic>=0.4.27",
    "chardet>=5.2",
    # Lightweight PDF text extraction (optional but pulled in for FastTextExtractor)
    "pypdf>=4.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]
# Layout-aware extraction via docling
layout = [
    "docling>=2.5",
]
# Vision extraction via OpenRouter GPT-4o-mini
# Also requires: poppler-utils (apt-get install poppler-utils)
vision = [
    "openai>=1.30",
    "pdf2image>=1.17",
    "Pillow>=10.3",
]
# Install everything at once
all = [
    "document-refinery[layout]",
    "document-refinery[vision]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
